Category functionality with user control
Category functionality with user control
