+++ IMS/Masters/Category.aspx
@@ -1,5 +1,7 @@
 ﻿<%@ Page Language="C#" MasterPageFile="~/Main.Master" AutoEventWireup="true" CodeBehind="Category.aspx.cs" Inherits="IMS.Category" %>
 
+<%@ Register src="~/UserControl/UC_Category.ascx" TagName="ctrlcategory"  TagPrefix="TWebControl"%>
+
 <%--<!DOCTYPE html>--%>
 <asp:Content ID="Content1" ContentPlaceHolderID="head" runat="server">
     
@@ -26,7 +28,7 @@
         }
     </script>
     <script type="text/javascript">
-        function CheckDouble() {
+       <%-- function CheckDouble() {
             $.ajax({
                 type: "POST",
                 url: '<%= ResolveUrl("~/Masters/Category.aspx/CheckDouble") %>', // this for calling the web method function in cs code.  
@@ -60,13 +62,13 @@
         function clearcategory() {
             cleartextboxes();
             $("#<%=txtCategoryName.ClientID%>").focus();
-        }
+        }--%>
     </script>
     <script type="text/javascript">
         // Reference the textbox and call its focus function
-        var txt = $("<%=txtCategoryName.ClientID%>");
+      <%--  var txt = $("<%=txtCategoryName.ClientID%>");
 
-        txt.focus();
+        txt.focus();--%>
     </script>
 </asp:Content>
 
@@ -77,41 +79,20 @@
             <a href="../Master.aspx" id="bMaster" runat="server">
                 <img src="../assets/img/goback-5-w800.png" height="50" width="130" /></a>
         </div>
+
+        
+       
+         
+
         <div class="panel panel-default text-center">
-            <div class="panel-heading">
+            <div class="panel-heading text-left">
                 <h1>Category Master</h1>
             </div>
-            <div class="panel-body">
-                <div class="form-horizontal">
-                    <div class="col-md-12">
-                        <div class="col-md-5">
-                            <div class="form-group ">
-                                <div class="col-sm-5 leftpadd0">
-                                    <label class="control-label   ">
-                                        Category Name:<asp:Label ID="lblStar" runat="server" Text="*" ForeColor="Red"></asp:Label>                                             
-                                    </label>
-                                </div>
-                                <div class="col-sm-7">
-                                    <asp:TextBox ID="txtCategoryName" onchange="CheckDouble()" runat="server" CssClass="form-control"></asp:TextBox>
-                                    <asp:RequiredFieldValidator ID="rvaldt1" runat="server" Display="Dynamic" ErrorMessage="Category name is required" ValidationGroup="abc" ControlToValidate="txtCategoryName" ForeColor="Red"></asp:RequiredFieldValidator>
-                                    <asp:HiddenField ID="HiddenField1" runat="server" />
-                                </div>
-                            </div>
-                        </div>
-                        <div class="col-md-7">
-                            <asp:Label ID="lblcheckDoubleError" runat="server"></asp:Label>
-                            <asp:HiddenField ID="hd" runat="server" />
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="panel-footer">
-
-                <asp:Button ID="btnSave" runat="server" CssClass="btn btn-primary " ValidationGroup="abc" Text="Save" OnClick="btnSave_Click"/>
-                <asp:Button ID="btnUpdate" runat="server" CssClass="btn btn-primary" ValidationGroup="abc" Text="Update" OnClick="btnUpdate_Click" Visible="false" />
-                <input class="btn btn-primary " type="button" value="Clear" onclick="javascript: window.location = 'Category.aspx'" />
-            </div>
+            
+           
+           
         </div>
+        <twebcontrol:ctrlcategory  ID="ctrlcategory"  runat="server" />
         <div class="row">
             <div class="alert alert-success" id="divalert" runat="server" visible="false">
                 <a href="#" class="close" data-dismiss="alert" aria-label="close">&times;</a>
